flask_server/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Django==2.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.0
paho-mqtt==1.3.1
pandas==0.23.4
python-dateutil==2.7.3
pytz==2017.3
requests==2.19.1
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.10
urllib3==1.23
Werkzeug==0.14.1
